agora foi luis

+++ Lab1/Shell.c
@@ -1,8 +1,21 @@
-/* 	Primeiro trabalho de lab de SO. 
-	Alunos: Luís Felipe Corrêa Ortolan RA:759375
-			Roseval Malaquias 						*/
+/*///////////////////////////////////////////////////////
+// Trabalho 2 - Criação de um programa shell           //
+//                                                     //
+// Autores:                                            //
+// Nome: Roseval Donisete Malaquias Junior             //
+// RA: 758597                                          //
+// Turma: A                                            //
+//                                                     //
+// Nome: Luís Felipe Corrêa Ortolan                    //
+// RA: 759375                                          //
+// Turma: A                                            //
+//                                                     //
+// O principal objetivo do trabalho é entender como o  //
+// shell funciona internamente. Dessa forma, utilizar  //
+// as informações adquiridas para implementar versão   //
+// simplificada de um programa shell.                  //
+///////////////////////////////////////////////////////*/
 
-/* Problema no fg quando da ctrl+c ele mata todo mundo, problema em colocar comando invalido na lista dentro do progfunc. */
 #include <unistd.h>
 #include <sys/types.h>
 #include <dirent.h>
@@ -17,6 +30,10 @@
 #include <sys/stat.h>
 #include <fcntl.h>
 #include <signal.h>
+#include <stdio.h>
+#include <stdlib.h>
+
+extern char **environ; // Variavel externa que possui as variaveis de ambiente.
 
 extern int errno; // Variavel utilizada para guardar erros ocorridos.
 
@@ -55,6 +72,44 @@ char **divLinha(char *buffer, int *tam){
 
 	for(i = 1; i < cont; i ++){
 		aux = strtok(NULL," ");
+		listaParametros[i] = malloc( (strlen(aux)+1) * sizeof(char));
+		listaParametros[i] = aux;
+
+	}
+
+	*tam = cont; // Retorna através da variavel tam o tamanho do vetor de string.
+	return listaParametros; // Retorna o vetor de strings com todas as palavras do buffer.
+}
+
+/* Essa função é uma versão alterada do divLinha, utilizada para "cortar" a variavel path, para que 
+ * seja retornado um vetor de strings contendo todas as paths presentes. Tem como entrada um ponteiro
+ * para char representado o buffer onde vai estar a variavel de ambiente "$PATH" e o tamanho do vetor
+ * de strings gerados.*/ 
+char **divPath(char *buffer, int *tam){
+	char *aux; // Variavel auxiliar para armazenar palavras do buffer.
+	int cont = 0;
+	int i = 0;
+
+	char *copiaBuffer = malloc(sizeof(char) * (strlen(buffer)+1));
+	strcpy(copiaBuffer,buffer); // Copia o buffer para outra variavel para utilizar a funcao strtok
+
+	/* Conta a quantidade de palavras dentro do buffer. */
+	aux = strtok(buffer,":");
+	while(aux != NULL){
+		cont++;
+		aux = strtok(NULL,":");
+	}
+
+	/* Instancia o vetor de strings com o tamanho referente a quantidade de palavras. */
+	char **listaParametros = malloc(sizeof(char *) * (cont+1) );
+
+	/* Encontra todas as palavras e coloca dentro do vetor de strings. */
+	aux = strtok(copiaBuffer,":");
+	listaParametros[i] = malloc( (strlen(aux)+1) * sizeof(char));
+	listaParametros[i] = aux;
+
+	for(i = 1; i < cont; i ++){
+		aux = strtok(NULL,":");
 		listaParametros[i] = malloc( (strlen(aux)+1) * sizeof(char));
 		listaParametros[i] = aux;
 
@@ -168,19 +223,38 @@ void progFunc(char **listaPalavras, int tam){
 				    dup2(fd1,2); // Faz o stdout apontar para o arquivo aberto
 				    close(fd1);
 				}
-
-			int erro = execve(listaPalavras[0],listaPalavras,NULL);
-			if(erro == -1)
-				printf("erro encontrado: %s\n",strerror(errno));
+			signal(SIGINT,SIG_DFL); // Define como default e tratamento desses sinais para o processo filho.
+			signal(SIGTSTP,SIG_DFL);
+			setpgrp(); // Seta o group id do processo filho, como seu proprio pid.
+			int erro = execve(listaPalavras[0],listaPalavras,environ);
+			
+			listaPalavras[tam-1] = NULL; // Retira o campo que contém "&" para passar como parametro.
+			/* Caso o arquivo nao esteja no diretório atual, procurar no PATH. */
+			if(erro == -1){
+				char *aux = getenv("PATH"); // Guarda o PATH dentro de uma variavel auxiliar.
+				int tam;
+				char **listaPath = divPath(aux,&tam); // Gerado um vetor de strings com cada path.
+				int i = 0;
+				/* Tenta inicializar o programa, em todas as pathas disponiveis, caso nao consiga comando não encontrado. */
+				while(i < tam){
+					char paths[300];
+					strcpy(paths,listaPath[i]);
+					strcat(paths,"/");
+					strcat(paths,listaPalavras[0]);
+					execve(paths,listaPalavras,environ);
+					i++;
+				}
+			}
 			printf("miniShell: %s: comando não encontrado\n",listaPalavras[0]);
 			exit(0); // Preciso mandar uma mensagem para o processo pai antes de morrer para conseguir retornar o -1.
 		}
-		insere(&ini, status, listaPalavras, tam);
+		insere(&ini, status, listaPalavras, tam); // Insere na lista de processos e printa seu indice e pid.
 		Celula *processo = busca(ini,status);
 		printf("[%d]",processo->chave);
 		printf(" %d\n",processo->pid);
 		return;	
 	}
+
 	else{
 		/* Execução de programa com parametros foreground. */
 		status = fork();
@@ -244,18 +318,52 @@ void progFunc(char **listaPalavras, int tam){
 				    dup2(fd1,2); // Faz o stdout apontar para o arquivo aberto
 				    close(fd1);
 				}
-
-			int erro = execve(listaPalavras[0],listaPalavras,NULL);
-			if(erro == -1)
-				printf("erro encontrado: %s\n",strerror(errno));
+			signal(SIGINT,SIG_DFL); // Define como default e tratamento desses sinais para o processo filho.
+			signal(SIGTSTP,SIG_DFL);
+			setpgrp(); // Seta o group id do processo filho, como seu proprio pid.
+			int erro = execve(listaPalavras[0],listaPalavras,environ);
+			/* Caso o arquivo nao esteja no diretório atual, procurar no PATH. */
+			if(erro == -1){
+				char *aux = getenv("PATH"); // Guarda o PATH dentro de uma variavel auxiliar.
+				int tam;
+				char **listaPath = divPath(aux,&tam); // Gerado um vetor de strings com cada path.
+				int i = 0;
+				/* Tenta inicializar o programa, em todas as pathas disponiveis, caso nao consiga comando não encontrado. */
+				while(i < tam){
+					char paths[300];
+					strcpy(paths,listaPath[i]);
+					strcat(paths,"/");
+					strcat(paths,listaPalavras[0]);
+					execve(paths,listaPalavras,environ);
+					i++;
+				}
+			}
 			printf("miniShell: %s: comando não encontrado\n",listaPalavras[0]);
 			exit(0);
 		}
 		else{
-			pidForground = status;
-			waitpid(status,NULL,WUNTRACED);
-		}
+			/* Processo pia precisa designar o controlador do terminal para o processo filho. */
+			pidForground = status; // Define o pid do processo em forground.
+			int erro = tcsetpgrp(STDIN_FILENO,pidForground); // Designa o controlador do terminal.
+			if(erro != 0)
+				printf("erro encontrado: %s\n",strerror(errno));
+			waitpid(pidForground,&status,WUNTRACED); // Wait no processo.
+
+			/* Caso o processo tenha recebido um sinal de SIGSTSP. */
+			if(WIFSTOPPED(status)){ // Coloca na fila de processos com o estado parado.
+				Celula *p = NULL;
+
+				p = busca(ini,pidForground);
+				if(p == NULL){
+					insere(&ini, pidForground, ultimoComando, tamComando);
+					p = busca(ini,pidForground);
+				}
+				strcpy(p->estado, "Parado ");
+				printf("\n[%d] %s %s\n", p->chave, p->estado, p->comando); 
+			}
+
 		return;
+			}
 	}
 }
 
@@ -275,14 +383,13 @@ int cd(char *listaPalavras, int tam){
 			return 1;
 	}
 	FILE *file;
-	if(file = fopen(listaPalavras,"r"))
+	if((file = fopen(listaPalavras,"r")))
 		fclose(file);
 	else{
-		printf("Arquivo nao existe\n");
+		printf("miniShell: cd: Arquivo ou diretório inexistente\n");
 		return 1;
 	}
 	if(listaPalavras[0] != '/'){
-		char aux[101];
 		strcat(caminho,"/");
 		strcat(caminho,listaPalavras);
 
@@ -321,13 +428,15 @@ int pwd(int opcao){
 		}
 	}
 }
-/*Verifica se os processos em segundo plano já acabaram*/
+
+/*Essa função tem como finalidade a cada sinal de prompt, retirar processos ja concluidos da lista de 
+ * processos. */
 void processoFinalizado(){
-	Celula *p = ini;
+	Celula *p = ini; // Apontador para o inicio da lista.
 	int stat = 1;
 	while(p != NULL){
-		stat = waitpid(p->pid, NULL, WNOHANG);
-		if(stat != 0){
+		stat = waitpid(p->pid, NULL, WNOHANG); // Termina processos zombies.
+		if(stat != 0){ // Processo encontrado, altera seu estado para Concluido, printa na tela depois o retira da lista.
 			strcpy(p->estado, "Concluido");
 			printf("[%d] %s %s\n", p->chave, p->estado, p->comando);
 			ini = retira(ini, p->pid);
@@ -337,35 +446,48 @@ void processoFinalizado(){
 }
 
 /* Rotina de tratamento para o SIGCHLD, tem como entrada o numero do signal. Essa rotina era checar, por
- * processos filhos zombies e era devidamente termina-los. */
+ * processos filhos zombies e era devidamente termina-los. Como os tratamentos de background e forground
+ * estao sendo feitos fora da rotina do sigchld, essa funcao é usada apenas para terminar propriamente
+ * os processos. */
 void sigchld_rotina(int signum){
-	int status; // status de retorno do processo filho ao terminar.
-	pid_t pid;
-	for(int i = 0; i < 20; i++) // for de 0 a 20, para caso um processo filho termine dentro da rotina de tratamento do SIGCHLD.
-		pid = waitpid(-1,&status,WNOHANG); // WNOHANG checa por zombies.
+	while(waitpid(-1,0,WNOHANG) > 0); // Enquanto tiver processos zombies o retorno é maior que zero.
 }
 
-/* Rotina de tratamento para o SIGINT, tem como entrada o numero do signal. Essa rotina termina o processo em forground. */
-void sigint_rotina(int signum){
-	if(pidForground != -1){
-		int error = kill(pidForground, SIGKILL); // Envia o sinal SIGKILL para o processo em forground.
-		if(error != 0) // Caso encontrou um erro.
-			printf("erro encontrado ao matar processo (%d): %s\n",pidForground, strerror(errno));
-		printf("\n");
-	}
-	else{
-		printf("\n");
-	}
-}
+/* Essa funcao tem como finalidade imitar o comando fg para colocar processos em forground, tem como 
+ * entrada a lista de palavras contendo o comando e seu atributo e o tamanho dessa lista. */
+int fg(char **listaPalavras,int tam){ 
+	int status; // Guarda o valor de retorno do waitpid.
+	int pos = 1; // Guarda a posicao do processo na lista.
+	Celula *p; // Ponteiro para a lista.
 
-void sigtstp_rotina(int signum){
-	if(pidForground != -1){
+	if(tam == 1){ // Caso tamanho seja 1, apenas o comando fg foi dado, aplicar a funcao no ultimo elemento da lista.
+		Celula *p = ini;
+		if(p != NULL)
+			while(p->prox != NULL){
+				p = p->prox;
+				pos++;
+			}
+	}
+	else
+		pos = atoi(listaPalavras[1]);
+
+	p = selecao(ini,pos);
+
+
+	if(p == NULL){ // Caso p seja NULL processo nao existe.
+		printf("miniShell: fg: o trabalho %d não existe\n",pos);
+		return 1;
+	}
+
+	pidForground = p->pid; // Altera o pid do processo em forground.
+	int erro = tcsetpgrp(STDIN_FILENO,pidForground); // Designa o controlador do terminal para o processo em forground.
+	if(erro != 0)
+		printf("erro encontrado: %s\n",strerror(errno));
+	waitpid(pidForground,&status,WUNTRACED); // Aguarda o processo.
+
+	if(WIFSTOPPED(status)){ // Caso saiu do waitpid com sinal de parada, alterar o estado do processo e colocar na lista.
 		Celula *p = NULL;
-		int error = kill(pidForground,SIGTSTP);
-		if(error != 0){
-			printf("erro encontrado ao matar processo (%d): %s\n", pidForground, strerror(errno));
-			return;
-		}
+
 		p = busca(ini,pidForground);
 		if(p == NULL){
 			insere(&ini, pidForground, ultimoComando, tamComando);
@@ -374,27 +496,45 @@ void sigtstp_rotina(int signum){
 		strcpy(p->estado, "Parado ");
 		printf("\n[%d] %s %s\n", p->chave, p->estado, p->comando); 
 	}
-}
-
-int fg(char **listaPalavras){
-	pid_t pid;                  
-	int pos = atoi(listaPalavras[1]);
-	Celula *p = selecao(ini,pos);
-	pidForground = p->pid;
-	waitpid(p->pid,NULL,WUNTRACED);
 	return 1;
 }
 
-int bg(char ** listaPalavras){
-	Celula *p;
-	int pos = atoi(listaPalavras[1]);
-	p = selecao(ini, pos);
-	strcpy(p->estado, "Executando ");
-	kill(p->pid, SIGCONT);
+/* Essa funcao tem como finalidade imitar o comando bg para colocar processos em background, tem como 
+ * entrada a lista de palavras contendo o comando e seu atributo e o tamanho dessa lista. */
+int bg(char ** listaPalavras, int tam){ 
+	int pos = 1; // Guarda a posicao do processo na lista.
+	Celula *p; // Ponteiro para a lista.
+
+	if(tam == 1){ // Caso tamanho seja 1, apenas o comando bg foi dado, aplicar a funcao no ultimo elemento da lista.
+		Celula *p = ini;
+		if(p != NULL)
+			while(p->prox != NULL){
+				p = p->prox;
+				pos++;
+			}
+	}
+	else
+		pos = atoi(listaPalavras[1]);
+
+
+	p = selecao(ini,pos);
+	
+	if(p == NULL){ // Caso esteja nesse estado, comando bg desconsiderado.
+		printf("miniShell: bg: o trabalho %d já está em plano de fundo\n",pos);
+		return 1;
+	}
+
+	if(strcmp(p->estado,"Executando ")!= 0){ // Caso ele nao esteja em estado executando.
+		strcpy(p->estado, "Executando ");
+		kill(p->pid, SIGCONT);
+	}
+
 	return 1;
 }
 
-/* Essa função tem como finalidade identificar o comando. */ // A ideia é essa funcao meio que organizar a logica, ai tera uma função pra cada funcionalidade do shell.
+/* Essa função tem como finalidade funcionar como o interpretador de comandos passado pelo usuario para o shell, tem como entrada o
+ * próprio buffer de entrada e como saida um inteiro sinalizando se o shell deve ser interrompido ou nao. Essa função identifica o 
+ * comando e chama a rotina apropriada para tratar desse comando.  */
 int interComando(char *buffer){
 	processoFinalizado();
 
@@ -402,32 +542,32 @@ int interComando(char *buffer){
 	if(buffer[0] == '\0')
 		return 1;
 
-	int status;  // Guarda o status para auxiliar no valor de retorno da função.
 	char ** listaPalavras; // Guarda as palavras presentes no buffer.
 	int tam = 0; // Guarda o tamanho do vetor de strings.
-	listaPalavras = divLinha(buffer,&tam);
+	listaPalavras = divLinha(buffer,&tam); // Utiliza o buffer para criar um vetor de strings.
 	ultimoComando = listaPalavras;
 	tamComando = tam;
-	if(strcmp(listaPalavras[0],"pwd") == 0 && tam == 1){
+
+	if(strcmp(listaPalavras[0],"pwd") == 0 && tam == 1){ // Comando pwd.
 		return pwd(1);
 	}
-	if(strcmp(listaPalavras[0],"jobs") == 0 && tam == 1){
+	if(strcmp(listaPalavras[0],"jobs") == 0 && tam == 1){ // Comando jobs.
 		imprime(ini);
 		return 1;
 	}
-	if(strcmp(listaPalavras[0],"fg") == 0 && tam == 2){
-		return fg(listaPalavras);
+	if(strcmp(listaPalavras[0],"fg") == 0 && (tam == 1 || tam == 2)){ // Comando fg.
+		return fg(listaPalavras,tam);
 	}
-	if(strcmp(listaPalavras[0],"bg") == 0 && tam == 2){
-		return bg(listaPalavras);
+	if(strcmp(listaPalavras[0],"bg") == 0 && (tam == 1 || tam == 2)){ // Comando bg.
+		return bg(listaPalavras,tam);
 	}
-	if(strcmp(listaPalavras[0],"cd") == 0){
+	if(strcmp(listaPalavras[0],"cd") == 0){ // Comando cd.
 		return cd(listaPalavras[1], tam);
 	}
-	if(strcmp(listaPalavras[0],"exit") == 0 && tam == 1){
+	if(strcmp(listaPalavras[0],"exit") == 0 && tam == 1){ // Comando exit.
 		return 0;
 	}
-	if(strcmp(listaPalavras[0],"clear") == 0 && tam == 1){
+	if(strcmp(listaPalavras[0],"clear") == 0 && tam == 1){ // Comando clear.
 		printf("\e[1;1H\e[2J"); // Limpa o terminal.
 		return 1;
 	}
@@ -441,14 +581,17 @@ int interComando(char *buffer){
 }
 
 
-
+/* Função main, onde sera recebido a entrada do usuario e printado o sinal de prompt na tela. */
 int main(){
 
 	char buffer[255]; // Buffer para ler o comando.
-	memset(buffer,0,sizeof(buffer));
-	int resp;
-	signal(SIGTSTP,sigtstp_rotina); // Tratar SIGTSTP. esse aqui que vai ser usado para parar o processo em forground...
-	signal(SIGINT,sigint_rotina); // Tratar SIGINT. (ctrl+c)
+	memset(buffer,0,sizeof(buffer)); // Seta todas as posicoes com valor 0.
+	int resp; // Guarda o retorno do interComando.
+	int pidShell = getpid(); // Guarda o pid do miniShell.
+
+	signal(SIGTTOU, SIG_IGN);
+	signal(SIGTSTP,SIG_IGN); // Tratar SIGTSTP. esse aqui que vai ser usado para parar o processo em forground...
+	signal(SIGINT,SIG_IGN); // Tratar SIGINT. (ctrl+c)
 	signal(SIGCHLD,sigchld_rotina); // Tratar SIGCHLD.
 	printf("\e[1;1H\e[2J");// Limpa o terminal antes de inicilizar o miniShell.
 
@@ -462,8 +605,14 @@ int main(){
 		printf("\033[0m"); // Reseta a cor.
 		fgets(buffer,254,stdin); // Recebe a entrada em buffer.
 		buffer[strlen(buffer)-1] = '\0';
+
 		resp = interComando(buffer); // Interpreta a entrada
 		memset(buffer,0,sizeof(buffer)); // Reseta o buffer.
+
+		int erro = tcsetpgrp(STDIN_FILENO,pidShell); // Seta o processo em forground para o controlador como o miniShell.
+			if(erro != 0)
+		printf("erro encontrado: %s\n",strerror(errno));
+
 	}while(resp);
 
 	return 0;
